+++ Source/Globals.h
@@ -20,7 +20,7 @@ void log(const char file[], int line, const char* format, ...);
 #define PIXEL_X_METER 16.f
 #define METERS_X_PIXEL 1.f/PIXEL_X_METER
 
-#define METERS_TO_PIXELS(m) (PIXEL_X_METER * (m))
+#define METERS_TO_PIXELS(m) (int)(PIXEL_X_METER * (m))
 #define PIXELS_TO_METERS(p) (METERS_X_PIXEL * (p))
 
 typedef unsigned int uint;
